Add dockerfile for chrome extension builds
