+++ chrome_extension/docker-compose.yml
@@ -1,0 +1,12 @@
+version: "3.7"
+services:
+  node:
+    build:
+      dockerfile: Dockerfile
+    volumes:
+      - chrome-extension-node-modules:/app/node_modules
+      - .:/app
+    ports:
+      - "3001:3001"
+volumes:
+  chrome-extension-node-modules:
